extract formatted CSV output function from samples and use it for grid search
